Rearrange storage package
